chore: merge configs folders
